cleanup code, added coments for all functions, fixed warnings in build

+++ OccurrenceMatrix.cpp
@@ -7,7 +7,12 @@ std::unordered_map<char, int> Precompute_DNATab = {
 	{'T', 3}
 };
 
-// Function to fill the matrix with occurrences of letters at each position in words
+/// <summary>
+/// Generates an occurrence matrix for DNA sequences.
+/// </summary>
+/// <param name="sequence">The DNA sequence to analyze.</param>
+/// <param name="word_size">The size of the word to split the sequence into.</param>
+/// <returns>A 4 x word_size matrix representing nucleotide occurrences.</returns>
 std::vector<std::vector<int>> GenerateOccurrenceMatrix(std::string_view sequence, int word_size)
 {
 	// Initialize a 4 x word_size matrix with zeros
@@ -34,8 +39,12 @@ std::vector<std::vector<int>> GenerateOccurrenceMatrix(std::string_view sequence
 	return matrix;
 }
 
-
-// Function to sum two matrices
+/// <summary>
+/// Adds two matrices element-wise.
+/// </summary>
+/// <param name="matrixA">The first matrix.</param>
+/// <param name="matrixB">The second matrix to add.</param>
+/// <returns>The resulting matrix after addition.</returns>
 std::vector<std::vector<int>> sumMatrices(const std::vector<std::vector<int>>& matrixA,
 	const std::vector<std::vector<int>>& matrixB)
 {
@@ -60,7 +69,12 @@ std::vector<std::vector<int>> sumMatrices(const std::vector<std::vector<int>>& m
 	return result; // Return the resulting matrix
 }
 
-// Function to subtract two matrices
+/// <summary>
+/// Subtracts two matrices element-wise.
+/// </summary>
+/// <param name="matrixA">The first matrix.</param>
+/// <param name="matrixB">The matrix to subtract.</param>
+/// <returns>The resulting matrix after subtraction.</returns>
 std::vector<std::vector<int>> subtractMatrices(const std::vector<std::vector<int>>& matrixA,
 	const std::vector<std::vector<int>>& matrixB)
 {
@@ -85,7 +99,13 @@ std::vector<std::vector<int>> subtractMatrices(const std::vector<std::vector<int
 	return result; // Return the resulting matrix
 }
 
-
+/// <summary>
+/// Adds a new DNA sequence's occurrence matrix to an existing one.
+/// </summary>
+/// <param name="prevMatrix">The previous occurrence matrix.</param>
+/// <param name="sequence">The new DNA sequence to add.</param>
+/// <param name="wordSize">The word size for the matrix.</param>
+/// <returns>The updated occurrence matrix.</returns>
 vector<vector<int>> AddSequenceToOccurrenceMatrix(const std::vector<std::vector<int>>& prevMatrix, std::string_view sequence, int wordSize)
 {
 	auto matrixToAdd = GenerateOccurrenceMatrix(sequence, wordSize);
@@ -94,6 +114,13 @@ vector<vector<int>> AddSequenceToOccurrenceMatrix(const std::vector<std::vector<
 	return matrixAfterAdd;
 }
 
+/// <summary>
+/// Removes a DNA sequence's occurrence matrix from an existing one.
+/// </summary>
+/// <param name="prevMatrix">The previous occurrence matrix.</param>
+/// <param name="sequence">The DNA sequence to remove.</param>
+/// <param name="wordSize">The word size for the matrix.</param>
+/// <returns>The updated occurrence matrix after removal.</returns>
 vector<vector<int>> RemoveSequenceFromOccurrenceMatrix(const std::vector<std::vector<int>>& prevMatrix, std::string_view sequence, int wordSize)
 {
 	auto matrixToSub = GenerateOccurrenceMatrix(sequence, wordSize);
@@ -102,11 +129,17 @@ vector<vector<int>> RemoveSequenceFromOccurrenceMatrix(const std::vector<std::ve
 	return matrixAfterSub;
 }
 
+/// <summary>
+/// Calculates the total percentage sum of the maximum nucleotide occurrences per column
+/// and constructs the representative word based on highest frequency nucleotides.
+/// </summary>
+/// <param name="matrix">The occurrence matrix.</param>
+/// <returns>A pair consisting of the total percentage sum and the representative word.</returns>
 std::pair<double, std::string> CalculatePercentageSumAndWord(const std::vector<std::vector<int>>& matrix)
 {
 	double totalSum = 0.0;           // Variable to store the total percentage sum
 	std::string representativeWord(matrix[0].size(), ' ');            // String to store the word with the best scores
-	int numColumns = matrix[0].size(); // Number of columns (word size)
+	size_t  numColumns = matrix[0].size(); // Number of columns (word size)
 
 	char DNATabReverse[4] = { 0 };
 	DNATabReverse[0] = 'A';
